Modificado mapping de Distancia Total y Pasos Totales por Int64

+++ Kilometros Database/UserDataTotalDistanceView.cs
@@ -16,8 +16,8 @@ namespace KilometrosDatabase
     {
         public System.Guid User_Guid { get; set; }
         public Nullable<System.DateTime> Timestamp { get; set; }
-        public Nullable<int> RunningDistance { get; set; }
-        public Nullable<int> WalkingDistance { get; set; }
+        public Nullable<long> RunningDistance { get; set; }
+        public Nullable<long> WalkingDistance { get; set; }
         public Nullable<int> TotalDistance { get; set; }
     }
 }
